New uart_reader base
